# api/config/swaggerConfig.yaml
openapi: 3.0.0
info:
  title: Jurassic-Park API
  version: "1.0.0"
  description: REST API for jurassic park Incident
servers:
  - url: http://127.0.0.1:3000
paths:
  /zones:
    get:
      summary: Liste toute les zones
      tags:
        - Zones
      responses:
        '200':
          description: Liste des zones
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Zone'

    post:
      summary: créer une zone
      tags:
        - Zones
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ZoneInput'
      responses:
        '200':
          description: Création d'une zone
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Zone'

  /zones/{id}:
    get:
      summary: Liste zone
      tags:
        - Zones
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Liste zone par id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Zone'

    patch:
      summary: modifier une zone
      tags:
        - Zones
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ZoneInput'
      responses:
        '200':
          description: Modification d'une zone
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

    delete:
      summary: Supprimer une zone
      tags:
        - Zones
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Supprimer une zone
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

  /incidents:
    get:
      summary: Liste tout les incidents
      tags:
        - Incidents
      responses:
        '200':
          description: Liste des incidents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Incidents'

    post:
      summary: créer un incident
      tags:
        - Incidents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentInput'
      responses:
        '200':
          description: Création d'un incident
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incidents'

  /incidents/{id}:
    get:
      summary: Liste incident
      tags:
        - Incidents
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Liste incident par id
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Incident'

    patch:
      summary: modifier un incident
      tags:
        - Incidents
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IncidentModification'
      responses:
        '200':
          description: Modification d'un incident
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

    delete:
      summary: Supprimer incident
      tags:
        - Incidents
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Supprimer un incident
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'

  /incidents/zone/{zoneId}:
    get:
      summary: Liste incident par zone
      tags:
        - Incidents
      parameters:
        - in: path
          name: zoneId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Liste incident filtrer par zone
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Incidents'

  /incidents/type/{type}:
    get:
      summary: Liste incident par type
      tags:
        - Incidents
      parameters:
        - in: path
          name: type
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Liste incident filtrer par type
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Incidents'

  /incidents/emergencyLevel/{emergencyLevel}:
    get:
      summary: Liste incident par urgence
      tags:
        - Incidents
      parameters:
        - in: path
          name: emergencyLevel
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Liste incident filtrer par urgence
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Incidents'

tags:
  - name: Incidents
    description: Geré les incidents
  - name: Zones
    description: Geré les zones

components:
  schemas:
    Incidents:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Hollogramme HS
        description:
          type: string
          example: description....
        type:
          type: string
          example: Panne
        zoneId:
          type: integer
          example: 1
        emergencyLevel:
          type: string
          example: moyenne
        status:
          type: boolean
          example: true

    Incident:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Hollogramme HS
        description:
          type: string
          example: description....
        type:
          type: string
          example: Panne
        zoneId:
          type: integer
          example: 1
        zone:
          type: object
          properties:
            id:
              type: integer
              example: 1
            name:
              type: string
              example: T-Rex valley
        emergencyLevel:
          type: string
          example: moyenne
        status:
          type: boolean
          example: true

    IncidentInput:
      type: object
      required:
        - title
        - description
        - type
        - zoneId
        - emergencyLevel
        - status
      properties:
        title:
          type: string
          example: Hollogramme HS
        description:
          type: string
          example: description....
        type:
          type: string
          example: Panne
        zoneId:
          type: integer
          example: 1
        emergencyLevel:
          type: string
          example: moyenne
        status:
          type: boolean
          example: true

    IncidentModification:
      type: object
      properties:
        title:
          type: string
          example: Hollogramme HS
        description:
          type: string
          example: description....
        type:
          type: string
          example: Panne
        zoneId:
          type: integer
          example: 1
        emergencyLevel:
          type: string
          example: moyenne
        status:
          type: boolean
          example: true

    Zone:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: T-Rex valley

    ZoneInput:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: T-Rex valley

    Message:
      type: object
      properties:
        message:
          type: string
          example: L'incident a été modifier avec succès